+++ spanner_iface.go
@@ -1,0 +1,5 @@
+package session
+
+type PreAuthHandlers interface { // only including this to avoid having to export the sessionHandlers interface
+	sessionHandlers
+}
